[LINT] error 578
Declaration of symbol 'Symbol' hides symbol 'Symbol'

Usage of `using namespace` muddied the global namespace with a lot of names
that had a very real potential for colliding. Only because cl.exe (VS's
compiler) tank-like abilities to compile anything was this working in some
cases.

There are still some 'error 578' left, but we'll take care of that in a later
commit.

Signed-off-by: Jocelyn Legault <jocelynlegault@gmail.com>