Revert "Remove useless casts."
Sadly, the casts weren't useless. :)
str() has a constructor only for const char *, and object(), so
char * is attempted to be converted to an object first, and there is
no converter for that.

An option would be add a converter, but I think that would be slower than
using the const char * constructor from str()

This reverts commit 37da9c9d7f20ee9e4f5cd58dda672f43724761b8.