Fix the const_cast problem.
The boost::python::str constructor accepts only a const char* and a non-const
simply doesn't cut it. However, I felt putting const_cast's everywhere in the
code was just plain ugly. This is the solution I came up with.

I've also normalized the spacing between the functions in the python script.
It's not made to offend, but I guess my the OCD in me felt very uncomfortable
dealing with that.

Signed-off-by: Jocelyn Legault <jocelynlegault@gmail.com>