[LINT] error 732
Loss of sign (Context) (Type to Type)

This mostly consists of a change to unsigned integers for any values
pertaining to sizes and indices. Sizes have been changed to size_t and indices
to idx_t (newly defined in stdafx.h).

However, the changes are rather invasive and I'm not 100% about their added
values. Should the maintainer of the project feel this commit isn't worthy of
the project, I'd understand if it's cherry picked out of the mainline branch.

Signed-off-by: Jocelyn Legault <jocelynlegault@gmail.com>