[LINT] error 730
Boolean argument to function.

The actual error states that we should not be using booleans as argument to a
function, but rather enums, as the meaning when calling a function is clearer,
for example: The meaning of DeleteDir(true,true); is much less understandable
than DeleteDir(DeleteReadOnly,RecurseSubfolders);
However, in the present case, we're dealing with Win32 API, and what jumped to
evidence was that built-in booleans (bool, true, false) were used in place of
the defined BOOL pseudo-type (TRUE, FALSE) used by the API. They're still
booleans passed to functions, but that's Win32 for you.

Signed-off-by: Jocelyn Legault <jocelynlegault@gmail.com>