Bugfix for pymlreplace
The offset had been converted to an idx_t, but in some circumstances it
needed to be negative.  Introduced an offset_t (int) typedef, and made
the currentOffset an offset_t.  This probably breaks some lint rules about
signed/unsigned conversions, but not sure what is best to do.

Also fixed a logic error with the "next" position to start from after
making a replacement.  The Python version of the string obviously doesn't
change after a replacement, therefore the next point can be immediately after
the end of the previous found target, ignoring the size of the replacement
string.