Fix for extract returning a temporary object
extract<T> returns a temporary object that converts to T.
Previously we were using this object further in the code.
This is now copied to a std::string, with the length included, so nulls should be included.

Note that NULL characters are currently treated as terminators for the search and replace strings by the Replacer.  This can (and should) be repaired later by providing the length in the call.