GIL Manager introduced. Lots of debugging info
Still not right. From the logs it appears the runScript thread reacquires the GIL at the end of add Text,
updates the thread ID in the GILManager, then the UI thread tries to acquire the GIL, *and get's it*, which
can't possibly be the case, unless something is wrong with the GIL calls I'm making.

See the log in notes/Thread-Debug.log
PY>[theadid]   [line number] [text]

Obviously the order of the line number is not definitely the order in which things happened,
but the "after" events *must* be after the event.