Deadlock solved: replace original thread ID when releasing GIL
Python pulses the GIL inside its main run-loop. What was happening
was that if another thread was waiting on the GIL (eg. our callback consumer)
then that thread stole the GIL, did it's thing, then released it. The pulse
then reclaims the GIL, and execution continues. Problem is then that the
m_currentThreadWithGIL was then reset to 0, as it was released again from
the thread that "stole" it (ie. the callback consumer).

The fix is simple - when taking the GIL, just record which thread had it before.
It will be 0 in the normal case, but in case we've stolen the GIL from
another thread inside a Python exec-loop, it will be the value of the thread
currently running Python code.  When we're done and release the GIL, we just
put it back to its original value, and Python execution grabs the GIL back
and continues as normal.  If our code in this original thread wants to perform
a temporary release, the m_currentThreadWithGIL is correct, and so it knows
to release it.