GILLock / GILRelease check if they have the GIL with API
In the case of 3 threads fighting over the GIL, it was possible for a deadlock
to occur, the thread 2 stole the GIL from thread 1, and thread 3 stole from
thread 2.

The API call to check if you have the GIL was added as a helper function
in Python 3.4.0 (issue #17522).  The function is taken from there.  A backport
would be trivial, but that would mean we could only use the packaged
python27.dll, and not the official one.