Generated test cases for ScintillaWrapper
Started implementation, and fixed some issues
We *always* need to release the GIL when we call Scintilla, even if it's just a read.
Something can fire (UPDATEUI), and we are then stuck on the main thread
waiting for the GIL.  If there's *any* callbacks registered, we can deadlock,
even if the notification fired hasn't got a callback, as we need the GIL
to check what callbacks are registered!

Each unique combination of return-type, param1Type, param2Type had a
test case and a callback test case generated. Implemented up to
void_int_colour.