Scintilla sync callbacks
editor.callbackSync() adds a synchronous callback for Scintilla.
Scintilla (at least as of version 3.3.9) runs the callbacks iterating on a
vector::iterator. If the callback causes a change in the document watchers
vectors then the iterator is no longer valid.  The BoostRegexSearch uses
a document watcher have a nice replace UX, so searchInTarget and findText
are affected. Any changing of doc pointers or creating new documents also
causes changes to the watchers.

To cope with this, these operations are explicitly disallowed in synchronous
callbacks. DepthCounter is used to keep track of if we are currently in a
callback or not.