Always release the GIL when calling N++
Notifications can come at any time due to WM_NOTIFY messages from child windows being forwarded by N++ to plugins. This means we can always get a notification, even when we're not expecting it.

Added check to see if we're interested in the notification before we claim the GIL. This is a belt-and-braces fix, as actually, simply performing a check if we're interested in the notification would be enough. However, there are so many hidden cases, always giving up the GIL whilst calling into N++ will eliminate all chance of deadlocks, and we can optimize later if it becomes an issue.

Reported by "Kadner"
https://sourceforge.net/p/npppythonscript/discussion/1188886/thread/915d6d40/?limit=25